add asserts to the test and commented exact diagonalization
