+++ applications/chemistry/projection_based_embedding/projected_based_embedding_tutorial.metadata.json
@@ -1,0 +1,10 @@
+{
+  "title": "Projected Based Embedding Tutorial",
+  "subtitle": "Embedding Hamiltonian Tutorial",
+  "description": "Embedding Hamiltonian Tutorial",
+  "friendly_name": "Projected Based Embedding Tutorial",
+  "vertical_tags": [],
+  "problem_domain_tags": [],
+  "qmod_type": [],
+  "level": []
+}
